refactor(connections): split XML-RPC client into package, extract JSON/2 ORM mixin
- odoo_connection.py (1073) -> odoo_connection/{core,auth,orm}.py with an
  __init__ facade re-exporting OdooConnection, create_connection, and
  OdooConnectionError
- odoo_json2_connection.py (586) -> transport stays put (incl. the
  cffi_requests patch target), ORM wrappers move to odoo_json2_orm.py
  as Json2OrmMixin

Both clients still satisfy OdooConnectionProtocol and stay separate per
the repo convention.
